need validate

+++ src/db/queries/grocerylistitem.js
@@ -1,3 +1,4 @@
+const Group = require("../models").Group;
 const User = require("../models").User;
 const GroceryListItem = require("../models").GroceryListItem;
 const GroceryList = require("../models").GroceryList;
@@ -14,6 +15,7 @@ module.exports = {
       callback(null, groceryListItem);
     })
     .catch((err) => {
+
       console.log(err);
       console.log("an err");
       callback(err);
@@ -34,49 +36,68 @@ module.exports = {
         callback(null, groceryListItem);
       })
       .catch((err) => {
-        console.log(err);
         callback(err);
       })
   },
 
   deleteGroceryListItem(id, callback){
-     return GroceryListItem.destroy({
-       where: {id}
-     })
-     .then((groceryListItem) => {
-       callback(null, groceryListItem);
-     })
-     .catch((err) => {
-       callback(err);
-     })
+    GroceryListItem.findById(id)
+    .then((data)=>{
+      if(!data){
+        let msg = {"success":false,"error" : "item not found"};
+        return callback(null,msg);
+      }
+        GroceryListItem.destroy({
+         where: {id}
+       })
+       .then((groceryListItem) => {
+         let msg = {"success":true};
+         return callback(null,msg);
+       })
+       .catch((err) => {
+         callback(err);
+       })
+    })
+
    },
 
   getGroceryListItem(id, callback){
     return GroceryListItem.findById(id, {
       include: [
-         {model: GroceryList},
-         {model: User}
+         {model: GroceryList, as:"grocerylist",
+            include: [
+              {model: User, as:"owner" },
+              {model: Group, as:"group" }
+            ]
+         },
+         {model: User, as:"buyer"}
        ]
      })
      .then((groceryListItem) => {
+       if(!groceryListItem){
+         let msg = {"success":false,"error" : "item not found"};
+         return callback(null,msg);
+       }
        callback(null, groceryListItem);
      })
      .catch((err) => {
-       console.log(err);
+      // console.log(err);
        callback(err);
      })
   },
   updateGroceryListItem(id, updatedGroceryListItem, callback) {
     return GroceryListItem.findById(id)
     .then((groceryListItem) => {
+    //    console.log(updatedGroceryListItem.name);
         if(!groceryListItem){
-          return callback("GroceryListItem not found");
+          let msg = {"success":false,"error" : "item not found"};
+          return callback(null,msg);
         }
         groceryListItem.update(updatedGroceryListItem, {
           fields: Object.keys(updatedGroceryListItem)
         })
-        .then(() => {
-          callback(null, groceryListItem);
+        .then((data) => {
+          callback(null, data);
         })
         .catch((err) => {
           callback(err);
