+++ src/db/queries/group.js
@@ -1,6 +1,7 @@
 const User = require("../models").User;
 const GroupUser = require("../models").GroupUser;
 const Group = require("../models").Group;
+const GroceryList = require("../models").GroceryList;
 
 
 module.exports = {
@@ -27,32 +28,50 @@ module.exports = {
         callback(null, group);
       })
       .catch((err) => {
-        console.log(err);
+      //  console.log(err);
         callback(err);
       })
   },
 
   deleteGroup(id, callback){
-     return Group.destroy({
-       where: {id}
+     Group.findById(id)
+     .then((group)=>{
+       if(!group){
+         let msg = {"success":false,"error" : "group not found"};
+         return callback(null, msg);
+       }
+       Group.destroy({
+         where: {id}
+       })
+       .then((group) => {
+         let msg = {"success":true};
+         return callback(null, msg);
+       })
+       .catch((err) => {
+         callback(err);
+       })
      })
-     .then((group) => {
-       callback(null, group);
-     })
-     .catch((err) => {
+     .catch((err)=>{
        callback(err);
      })
+
+
    },
 
   getGroup(id, callback){
     return Group.findById(id, {
       include: [
          {
-           model: GroupUser, as: "groupusers", include: [{model: User }]
-         }
+           model: GroupUser, as: "groupusers", include: [{model: User, as:"user" }]
+         },
+         {model:GroceryList, as: "grocerylists"}
        ]
      })
      .then((group) => {
+       if(!group){
+         let msg = {"success":false,"error" : "group not found"};
+         return callback(null, msg);
+       }
        callback(null, group);
      })
      .catch((err) => {
@@ -63,13 +82,14 @@ module.exports = {
     return Group.findById(id)
     .then((group) => {
         if(!group){
-          return callback("Group not found");
+          let msg = {"success":false,"error" : "group not found"};
+          return callback(null, msg);
         }
         group.update(updatedGroup, {
           fields: Object.keys(updatedGroup)
         })
-        .then(() => {
-          callback(null, group);
+        .then((updated) => {
+          callback(null, updated);
         })
         .catch((err) => {
           callback(err);
