+++ src/db/queries/user.js
@@ -1,6 +1,7 @@
 const User = require("../models").User;
 const GroupUser = require("../models").GroupUser;
 const Group = require("../models").Group;
+const GroceryList = require("../models").GroceryList;
 
 
 module.exports = {
@@ -22,12 +23,17 @@ module.exports = {
      return User.findById(id, {
        include: [
           {
-            model: GroupUser, as: "groupusers", include: [{model: Group }]
-          }
+            model: GroupUser, as: "groupusers", include: [{model: Group, as: "group" }]
+          },
+          {model: GroceryList, as : "grocerylists"}
         ]
     })
 
      .then((user) => {
+       if(!user){
+         let msg = {"success":false,"error" : "user not found"};
+         return callback(null, msg);
+       }
        callback(null, user);
      })
      .catch((err) => {
@@ -47,28 +53,43 @@ module.exports = {
         callback(null, user);
       })
       .catch((err) => {
-        console.log(err);
         callback(err);
       })
   },
 
   deleteUser(id, callback){
-     return User.destroy({
-       where: {id}
-     })
-     .then((user) => {
-       callback(null, user);
-     })
-     .catch((err) => {
-       callback(err);
-     })
+    User.findById(id)
+    .then((user)=>{
+      if(!user){
+        let msg = {"success":false,"error" : "user not found"};
+        return callback(null, msg);
+      }
+      User.destroy({
+         where: {id}
+       })
+       .then((user) => {
+
+           let msg = {"success":true};
+           return callback(null, msg);
+       })
+       .catch((err) => {
+         callback(err);
+       })
+
+    })
+    .catch((err)=>{
+      console.log(err);
+      callback(err);
+    })
+
    },
 
   updateUser(id, updatedUser, callback) {
     return User.findById(id)
     .then((user) => {
         if(!user){
-          return callback("User not found");
+          let msg = {"success":false,"error" : "user not found"};
+          return callback(null, msg);
         }
         user.update(updatedUser, {
           fields: Object.keys(updatedUser)
